Add Module:Player/Ext
